requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-fakerlibrary
robotframework-databaselibrary
psycopg2-binary
